+++ TelemetryCore/src/main/cpp/Telemetry/cFiles/mavlink2.c
@@ -6,7 +6,7 @@
 mavlink_status_t status;
 mavlink_message_t msg;
 
-void mavlink_read_v2(UAVTelemetryData *td,OriginData *originData,const uint8_t *data,const int data_length) {
+void mavlink_read_v2(UAVTelemetryData *td,OriginData *originData,const uint8_t *data,const size_t data_length) {
     for(int i=0; i<data_length; i++) {
         uint8_t c = data[i];
         if (mavlink_parse_char(MAVLINK_COMM_0, c, &msg, &status)) {
