+++ TelemetryCore/src/main/cpp/Telemetry/cFiles/mavlink2.h
@@ -5,6 +5,6 @@
 #include "mavlink_v2/common/mavlink.h"
 #include "UAVTelemetryData.h"
 
-void mavlink_read_v2(UAVTelemetryData *td,OriginData *originData,const uint8_t *data, const int data_length);
+void mavlink_read_v2(UAVTelemetryData *td,OriginData *originData,const uint8_t *data, const size_t data_length);
 
 #endif
